fix: add full stub implementation of the 1.x interface
To avoid crashing an application that just upgraded to the 2.0 version, we provide a stub implementation for all public methods of the 1.x interface. The returned promises wont resolve and callbacks wont be called, so it should behave as if the maps API is still being loaded when the error-message is logged.